Add UIHelper class
